Clear useless code
